Move tests & benches to separate files
